Agent: Split request handler into its own module
